// Frontend/HSPA/src/app/model/IPropertyBase.ts
export interface IpropertyBase {
    Id?:number;
    SellRent?: number,
    Name? : string;
    PType?: string;
    FType?: string;
    Price?: number;
    BHK?: number;
    BuiltArea?: number;
    City? : string;
    RTM?: number;
    Image? : string;
    RadyToMove?:boolean;
    estPossessionOn?:Date;
    CarpetArea?: number;
    FloorNo?: string;
    TotalFloors?: string;
    Age?: string;
    MainEntrance?: string;
    Gated?: boolean;
    Security?: number;
    Maintenance?: number;
    Description?: string;
    Address?: string;
    Address2?: string;
}